working with no persistence
now we’re accepting messages from Phoenix and
storing them in minimongo. Instead of saving to a
mongo collection we’re sending data to phoenix.

However, this is not persisting anyone yet, and if
a user enters a chatroom they won’t be able to
see the previous posts (before they entered).
We’ll fix next.